+++ example/example/settings.py
@@ -6,7 +6,10 @@
 #     https://docs.scrapy.org/en/latest/topics/settings.html
 #     https://docs.scrapy.org/en/latest/topics/downloader-middleware.html
 #     https://docs.scrapy.org/en/latest/topics/spider-middleware.html
+
 from example.autoextract import AutoextractProductProvider
+
+from scrapy_poet import ScrapyPoetRequestFingerprinter
 
 BOT_NAME = "example"
 
@@ -20,4 +23,12 @@ ROBOTSTXT_OBEY = True
 
 DOWNLOADER_MIDDLEWARES = {
     "scrapy_poet.InjectionMiddleware": 543,
+    "scrapy.downloadermiddlewares.stats.DownloaderStats": None,
+    "scrapy_poet.DownloaderStatsMiddleware": 850,
+}
+
+REQUEST_FINGERPRINTER_CLASS = ScrapyPoetRequestFingerprinter
+
+SPIDER_MIDDLEWARES = {
+    "scrapy_poet.RetryMiddleware": 275,
 }