+++ scrapy_poet/__init__.py
@@ -1,4 +1,5 @@
 from .api import DummyResponse, callback_for
-from .downloadermiddlewares import InjectionMiddleware
+from .downloadermiddlewares import DownloaderStatsMiddleware, InjectionMiddleware
 from .page_input_providers import HttpResponseProvider, PageObjectInputProvider
 from .spidermiddlewares import RetryMiddleware
+from ._request_fingerprinter import ScrapyPoetRequestFingerprinter