+++ scrapy_poet/utils/testing.py
@@ -2,17 +2,19 @@ import json
 from inspect import isasyncgenfunction
 from typing import Dict
 
-from scrapy import signals
+from scrapy import Spider, signals
 from scrapy.crawler import Crawler
 from scrapy.exceptions import CloseSpider
 from scrapy.settings import Settings
 from scrapy.utils.python import to_bytes
+from scrapy.utils.test import get_crawler as _get_crawler
 from twisted.internet import reactor
 from twisted.internet.defer import inlineCallbacks
 from twisted.internet.task import deferLater
 from twisted.web.resource import Resource
 from twisted.web.server import NOT_DONE_YET
 
+from scrapy_poet import ScrapyPoetRequestFingerprinter
 from scrapy_poet.utils.mockserver import MockServer
 
 
@@ -150,7 +152,21 @@ def crawl_single_item(
     return item, url, crawler
 
 
-def make_crawler(spider_cls, settings):
+def get_download_handler(crawler, schema):
+    return crawler.engine.downloader.handlers._get_handler(schema)
+
+
+def make_crawler(spider_cls, settings=None):
+    settings = settings or {}
+    if isinstance(settings, dict):
+        _settings = create_scrapy_settings()
+        _settings.update(settings)
+    else:
+        _settings = create_scrapy_settings()
+        for k, v in dict(settings).items():
+            _settings.set(k, v, priority=settings.getpriority(k))
+    settings = _settings
+
     if not getattr(spider_cls, "name", None):
 
         class Spider(spider_cls):
@@ -160,6 +176,33 @@ def make_crawler(spider_cls, settings):
         Spider.__module__ = spider_cls.__module__
         spider_cls = Spider
     return Crawler(spider_cls, settings)
+
+
+def setup_crawler_engine(crawler: Crawler):
+    """Run the crawl steps until engine setup, so that crawler.engine is not
+    None.
+    https://github.com/scrapy/scrapy/blob/8fbebfa943c3352f5ba49f46531a6ccdd0b52b60/scrapy/crawler.py#L116-L122
+    """
+
+    crawler.crawling = True
+    crawler.spider = crawler._create_spider()
+    crawler.engine = crawler._create_engine()
+
+    handler = get_download_handler(crawler, "https")
+    if hasattr(handler, "engine_started"):
+        handler.engine_started()
+
+
+class DummySpider(Spider):
+    name = "dummy"
+
+
+def get_crawler(settings=None, spider_cls=DummySpider, setup_engine=True):
+    settings = settings or {}
+    crawler = _get_crawler(settings_dict=settings, spidercls=spider_cls)
+    if setup_engine:
+        setup_crawler_engine(crawler)
+    return crawler
 
 
 class CollectorPipeline:
@@ -186,7 +229,7 @@ class InjectedDependenciesCollectorMiddleware:
         return response
 
 
-def create_scrapy_settings(request):
+def create_scrapy_settings():
     """Default scrapy-poet settings"""
     s = dict(
         # collect scraped items to crawler.spider.collected_items
@@ -197,6 +240,12 @@ def create_scrapy_settings(request):
             # collect injected dependencies to crawler.spider.collected_response_deps
             InjectedDependenciesCollectorMiddleware: 542,
             "scrapy_poet.InjectionMiddleware": 543,
+            "scrapy.downloadermiddlewares.stats.DownloaderStats": None,
+            "scrapy_poet.DownloaderStatsMiddleware": 850,
+        },
+        REQUEST_FINGERPRINTER_CLASS=ScrapyPoetRequestFingerprinter,
+        SPIDER_MIDDLEWARES={
+            "scrapy_poet.RetryMiddleware": 275,
         },
     )
     return Settings(s)