+++ tests/test_downloader.py
@@ -31,6 +31,7 @@ from scrapy_poet.utils.testing import (
     DelayedResource,
     EchoResource,
     StatusResource,
+    create_scrapy_settings,
     make_crawler,
 )
 
@@ -156,12 +157,6 @@ def test_additional_requests_success() -> None:
         class TestSpider(Spider):
             name = "test_spider"
 
-            custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
-            }
-
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
@@ -169,7 +164,7 @@ def test_additional_requests_success() -> None:
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
         yield crawler.crawl()
 
     assert items == [{"foo": "bar"}]
@@ -197,12 +192,6 @@ def test_additional_requests_bad_response() -> None:
         class TestSpider(Spider):
             name = "test_spider"
 
-            custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
-            }
-
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
@@ -210,7 +199,7 @@ def test_additional_requests_bad_response() -> None:
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
         yield crawler.crawl()
 
     assert items == [{"foo": "bar"}]
@@ -246,12 +235,6 @@ def test_additional_requests_connection_issue() -> None:
             class TestSpider(Spider):
                 name = "test_spider"
 
-                custom_settings = {
-                    "DOWNLOADER_MIDDLEWARES": {
-                        "scrapy_poet.InjectionMiddleware": 543,
-                    },
-                }
-
                 def start_requests(self):
                     yield Request(server.root_url, callback=self.parse)
 
@@ -259,7 +242,7 @@ def test_additional_requests_connection_issue() -> None:
                     item = await page.to_item()
                     items.append(item)
 
-            crawler = make_crawler(TestSpider, {})
+            crawler = make_crawler(TestSpider)
             yield crawler.crawl()
 
     assert items == [{"foo": "bar"}]
@@ -293,13 +276,6 @@ def test_additional_requests_ignored_request() -> None:
         class TestSpider(Spider):
             name = "test_spider"
 
-            custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    TestDownloaderMiddleware: 1,
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
-            }
-
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
@@ -307,7 +283,9 @@ def test_additional_requests_ignored_request() -> None:
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        settings = create_scrapy_settings()
+        settings["DOWNLOADER_MIDDLEWARES"][TestDownloaderMiddleware] = 1
+        crawler = make_crawler(TestSpider, settings)
         yield crawler.crawl()
 
     assert items == [{"exc": HttpError}]
@@ -352,13 +330,6 @@ def test_additional_requests_unhandled_downloader_middleware_exception() -> None
         class TestSpider(Spider):
             name = "test_spider"
 
-            custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    TestDownloaderMiddleware: 1,
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
-            }
-
             def start_requests(self):
                 yield Request(server.root_url, callback=self.parse)
 
@@ -366,7 +337,9 @@ def test_additional_requests_unhandled_downloader_middleware_exception() -> None
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        settings = create_scrapy_settings()
+        settings["DOWNLOADER_MIDDLEWARES"][TestDownloaderMiddleware] = 1
+        crawler = make_crawler(TestSpider)
         yield crawler.crawl()
 
     assert items == [{"exc": HttpError}]
@@ -404,12 +377,6 @@ def test_additional_requests_dont_filter() -> None:
         class TestSpider(Spider):
             name = "test_spider"
 
-            custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
-            }
-
             def start_requests(self):
                 yield Request(server.root_url, body=b"a", callback=self.parse)
                 yield Request(server.root_url, body=b"a", callback=self.parse)
@@ -418,7 +385,7 @@ def test_additional_requests_dont_filter() -> None:
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
         yield crawler.crawl()
 
     assert items == [{"a": "a"}]
@@ -476,9 +443,6 @@ def test_additional_requests_no_cb_deps() -> None:
             name = "test_spider"
 
             custom_settings = {
-                "DOWNLOADER_MIDDLEWARES": {
-                    "scrapy_poet.InjectionMiddleware": 543,
-                },
                 "SCRAPY_POET_PROVIDERS": {
                     BrowserResponseProvider: 1100,
                 },
@@ -491,7 +455,7 @@ def test_additional_requests_no_cb_deps() -> None:
                 item = await page.to_item()
                 items.append(item)
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
         yield crawler.crawl()
 
     assert provider_calls == 1
@@ -506,11 +470,6 @@ class BasicPage(WebPage):
 
 class BaseSpider(Spider):
     name = "test_spider"
-    custom_settings = {
-        "DOWNLOADER_MIDDLEWARES": {
-            "scrapy_poet.InjectionMiddleware": 543,
-        },
-    }
 
 
 # See: https://github.com/scrapinghub/scrapy-poet/issues/48
@@ -572,7 +531,7 @@ def test_parse_callback_none_dummy_response() -> None:
             def parse(self, response: DummyResponse):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -603,7 +562,7 @@ def test_parse_callback_none_response() -> None:
             def parse(self, response: scrapy.http.Response):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -618,8 +577,8 @@ def test_parse_callback_none_response() -> None:
 )
 @inlineCallbacks
 def test_parse_callback_none_no_annotated_deps() -> None:
-    """Similar to ``test_parse_callback_none_dummy_response()`` but there are no
-    annotated dependencies.
+    """Similar to ``test_parse_callback_none_dummy_response()`` but there are
+    no annotated dependencies.
 
     No warnings should be issued here.
     """
@@ -634,12 +593,12 @@ def test_parse_callback_none_no_annotated_deps() -> None:
             def parse(self, response):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
-        with warnings.catch_warnings(record=True) as warning_msg:
+        with warnings.catch_warnings(record=True) as w:
             yield crawler.crawl()
 
-    assert not warning_msg
+    assert not any("with callback=None" in str(warning.message) for warning in w)
     assert isinstance(collected["response"], scrapy.http.Response)
 
 
@@ -664,7 +623,7 @@ def test_parse_callback_none_with_deps(caplog) -> None:
             def parse(self, response: DummyResponse, page: BasicPage):
                 pass
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -707,7 +666,7 @@ def test_parse_callback_none_with_deps_cb_kwargs(caplog) -> None:
             def parse(self, response: DummyResponse, page: BasicPage):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -747,7 +706,7 @@ def test_parse_callback_none_with_deps_cb_kwargs_incomplete(caplog) -> None:
             ):
                 pass
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -784,7 +743,7 @@ def test_parse_callback_NO_CALLBACK(caplog) -> None:
             def parse(self, response: DummyResponse):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()
@@ -816,7 +775,7 @@ def test_parse_callback_NO_CALLBACK_with_page_dep(caplog) -> None:
             def parse(self, response: DummyResponse, page: BasicPage):
                 collected["response"] = response
 
-        crawler = make_crawler(TestSpider, {})
+        crawler = make_crawler(TestSpider)
 
         with pytest.warns(UserWarning) as record:
             yield crawler.crawl()